Author Service

+++ BookMarketRest/src/main/java/com/example/dto/requestDto/BookRequestDto.java
@@ -7,6 +7,6 @@ import java.util.List;
 @Data
 public class BookRequestDto {
     private String name;
-    private List<Long> authorIds;
-    private Long categoryId;
+    private List<Integer> authorIds;
+    private Integer categoryId;
 }
